Fix build errors by removing invalid setSession export from route.ts (#5)
